Move error icon in form field out of input container
