+++ packages/form-field/src/FormField/FormField.styles.ts
@@ -1,4 +1,6 @@
 import { css } from '@leafygreen-ui/emotion';
+import { Theme } from '@leafygreen-ui/lib';
+import { palette } from '@leafygreen-ui/palette';
 import { BaseFontSize, Size, spacing, typeScales } from '@leafygreen-ui/tokens';
 
 export const getFontSize = ({
@@ -39,6 +41,18 @@ export const marginBottom = css`
   margin-bottom: ${spacing[100]}px;
 `;
 
-export const errorTextContainerStyle = css`
+export const errorContainerStyle = css`
   margin-top: ${spacing[100]}px;
+  display: flex;
+  align-items: center;
+  gap: ${spacing[100]}px;
 `;
+
+export const errorIconStyles: Record<Theme, string> = {
+  [Theme.Light]: css`
+    color: ${palette.red.base};
+  `,
+  [Theme.Dark]: css`
+    color: ${palette.red.light1};
+  `,
+};
